Map -pi to pi for principal argument in complex operators with float values (bug #43313)
* Array-fC.cc (nan_ascending_compare, nan_descending_compare): Use static_cast
to float on M_PI so that comparisons work.

* oct-cmplx.h (DEF_COMPLEXR_COMP): Use static_cast to type <T> on M_PI so that
comparisons work.

* complex.tst: Add tests for single values on top of those for double.

+++ liboctave/array/Array-fC.cc
@@ -58,9 +58,10 @@ nan_ascending_compare (const FloatComplex& x, const FloatComplex& y)
     {
       float xarg = arg (x);
       float yarg = arg (y);
-      xarg = (xarg == -M_PI) ? M_PI : xarg;
-      yarg = (yarg == -M_PI) ? M_PI : yarg;
-
+      xarg = (xarg == -static_cast<float> (M_PI)) ? static_cast<float> (M_PI)
+                                                  : xarg;
+      yarg = (yarg == -static_cast<float> (M_PI)) ? static_cast<float> (M_PI)
+                                                  : yarg;
       return xarg < yarg;
     }
   else
@@ -82,9 +83,10 @@ nan_descending_compare (const FloatComplex& x, const FloatComplex& y)
     {
       float xarg = arg (x);
       float yarg = arg (y);
-      xarg = (xarg == -M_PI) ? M_PI : xarg;
-      yarg = (yarg == -M_PI) ? M_PI : yarg;
-
+      xarg = (xarg == -static_cast<float> (M_PI)) ? static_cast<float> (M_PI)
+                                                  : xarg;
+      yarg = (yarg == -static_cast<float> (M_PI)) ? static_cast<float> (M_PI)
+                                                  : yarg;
       return xarg > yarg;
     }
   else
